requirements.txt
gitdb2==2.0.6
GitPython==3.0.2
pytz==2019.2
smmap2==2.0.5
